Tunnel: better file management for tunnelproxymanager
