+++ go/src/koding/workers/tunnelproxymanager/sqs.go
@@ -1,0 +1,155 @@
+package main
+
+import (
+	"encoding/json"
+	"errors"
+	"fmt"
+
+	"github.com/aws/aws-sdk-go/aws"
+	"github.com/aws/aws-sdk-go/service/sqs"
+)
+
+// MakeSureSQS configures a queue for listening to a predefined system
+func (l *LifeCycle) MakeSureSQS(name string) error {
+	sqsLogger := l.log.New("SQS")
+	sqsLogger.Debug("working...")
+
+	queueName := "SQS-" + name
+	// create queue is idempotent, if it is already created returns existing one
+	// all Attributes should be same tho
+	createQueueResp, err := l.sqs.CreateQueue(&sqs.CreateQueueInput{
+		QueueName: aws.String(queueName), // Required
+		Attributes: map[string]*string{ // Required
+			// The time in seconds that the delivery of all messages in the
+			// queue will be delayed
+			"DelaySeconds": aws.String("0"),
+
+			//  The limit of how many bytes a message can contain
+			"MaximumMessageSize": aws.String("262144"), // 256 KiB
+
+			// The number of seconds Amazon SQS retains a message. Integer
+			// representing seconds, from 60 (1 minute) to 1209600 (14 days).
+			"MessageRetentionPeriod": aws.String("3600"), // 1 hour
+
+			//The time for which a ReceiveMessage call will wait for a message
+			//to arrive
+			"ReceiveMessageWaitTimeSeconds": aws.String("20"),
+
+			// The visibility timeout for the queue. An integer from 0 to 43200
+			// (12 hours). Messages will be available again if they are not
+			// deleted in VisibilityTimeout
+			"VisibilityTimeout": aws.String("3600"), // 60 secs
+		},
+	})
+
+	if err != nil {
+		return err
+	}
+
+	l.queueURL = createQueueResp.QueueURL
+
+	sqsLogger.Debug("Queue is created URL: %s", *l.queueURL)
+	sqsLogger.Debug("Configuring Queue...")
+
+	resp, err := l.sqs.GetQueueAttributes(&sqs.GetQueueAttributesInput{
+		QueueURL: l.queueURL, // Required
+		AttributeNames: []*string{
+			aws.String("QueueArn"),
+			aws.String("Policy"),
+		},
+	})
+	if err != nil {
+		return err
+	}
+
+	if resp == nil {
+		return errors.New("malformed response for GetQueueAttributes")
+	}
+
+	queueARN, ok := resp.Attributes["QueueArn"]
+	if !ok || queueARN == nil {
+		return errors.New("QueueArn not exists in Attributes")
+	}
+	l.queueARN = queueARN
+
+	p1 := newDefaultPolicy(*l.topicARN, *l.queueARN)
+	b, err := json.Marshal(p1)
+	if err != nil {
+		return err
+	}
+
+	// Attributes: {
+	//   QueueArn: "arn:aws:sqs:us-east-1:616271189586:SQS-ElasticBeanstalkNotifications-Environment-cihangir",
+	//   Policy: "{\"Version\":\"2012-10-17\",\"Id\":\"arn:aws:sqs:us-east-1:616271189586:SQS-ElasticBeanstalkNotifications-Environment-cihangir/SQSDefaultPolicy\",\"Statement\":[{\"Sid\":\"tunnelproxy_dev_1\",\"Effect\":\"Allow\",\"Principal\":{\"AWS\":\"*\"},\"Action\":\"SQS:SendMessage\",\"Resource\":\"arn:aws:sqs:us-east-1:616271189586:SQS-ElasticBeanstalkNotifications-Environment-cihangir\",\"Condition\":{\"ArnEquals\":{\"aws:SourceArn\":\"arn:aws:sns:us-east-1:616271189586:tunnelproxymanager_test\"}}}]}"
+	// }
+	if !isValid(resp.Attributes, string(b)) {
+		sqsLogger.Debug("Queue Policy is not correct, fixing it...")
+		_, err := l.sqs.SetQueueAttributes(&sqs.SetQueueAttributesInput{
+			Attributes: map[string]*string{ // Required
+				"Policy": aws.String(string(b)), // Required
+			},
+			QueueURL: l.queueURL,
+		})
+		if err != nil {
+			return err
+		}
+		sqsLogger.Debug("Queue Policy configured properly")
+	}
+
+	sqsLogger.Debug("Queue is ready")
+	return nil
+}
+
+type policy struct {
+	Version   string      `json:"Version"`
+	ID        string      `json:"Id"`
+	Statement []statement `json:"Statement"`
+}
+
+type statement struct {
+	Sid       string                 `json:"Sid"`
+	Effect    string                 `json:"Effect"`
+	Principal map[string]interface{} `json:"Principal"`
+	Action    string                 `json:"Action"`
+	Resource  string                 `json:"Resource"`
+	Condition map[string]interface{} `json:"Condition"`
+}
+
+func newDefaultPolicy(topicARN, queueARN string) *policy {
+	return &policy{
+		Version: "2012-10-17",
+		ID:      fmt.Sprintf("%s/SQSDefaultPolicy", queueARN),
+		Statement: []statement{
+			statement{
+				Sid:    callerReferance,
+				Effect: "Allow",
+				Principal: map[string]interface{}{
+					"AWS": "*",
+				},
+				Action:   "SQS:SendMessage",
+				Resource: queueARN,
+				Condition: map[string]interface{}{
+					"ArnEquals": map[string]string{
+						"aws:SourceArn": topicARN,
+					},
+				},
+			},
+		},
+	}
+}
+
+func isValid(attr map[string]*string, b string) bool {
+	if attr == nil {
+		return false
+	}
+
+	_, ok := attr["Policy"]
+	if !ok {
+		return false
+	}
+
+	if b != *attr["Policy"] {
+		return false
+	}
+	return true
+}
